cw1

+++ 1.2.8/main.c
@@ -5,6 +5,7 @@ int main()
 {
     int a, b, c;
     scanf("%d %d %d", &a, &b, &c);
-    printf("%f", (float)(a+b+c)/3);
+    //printf("%f", (float)(a+b+c)/3);
+    printf("%f", (a+b+c)/3.0);
     return 0;
 }
